# specs/001-gs-control-center/contracts/api.yaml
openapi: 3.1.0
info:
  title: GS Control Center API
  description: Centralized Advertising Block Management System for UCS Guest Screen
  version: 1.0.0
servers:
  - url: /api/v1
    description: Central Server API (Host 10.0.0.111)
paths:
...
  /publications:
    post:
      summary: Initiate or schedule a publication batch for an advertising block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [advertising_block_id, scope_type, scope_target_ids]
              properties:
                advertising_block_id: { type: string, format: uuid }
                scope_type: { type: string, enum: [REGION, BRANCH, CUSTOM_CASHIERS] }
                scope_target_ids:
                  type: array
                  items: { type: string, format: uuid }
                scheduled_at: { type: string, format: date-time, nullable: true }
      responses:
        '202':
          description: Publication batch accepted and enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationBatch'
  /publications/{id}:
    get:
      summary: Get publication batch progress and per-cashier status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationBatchDetail'
  /publications/{id}/retry:
    post:
      summary: Retry failed or offline cashiers within a batch
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: Retry jobs queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  retried_count: { type: integer }
  /publications/{id}/retry-failed:
    post:
      summary: Retry strictly failed or offline cashiers within a batch using original immutable snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: Retry jobs dispatched strictly for failed/offline nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  retried_count: { type: integer }
                  batch_id: { type: string, format: uuid }
...
  /events/publications/{batch_id}:
    get:
      summary: Real-time SSE stream of publication progress events
      parameters:
        - name: batch_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Event stream (Server-Sent Events)
          content:
            text/event-stream:
              schema:
                type: string
                description: JSON events formatted as data:<payload>\n\n
...
components:
  schemas:
...
    PublicationBatchSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        advertising_block_name: { type: string }
        area: { type: string, enum: [FULL_SCREEN, MODE32_PROMO] }
        status: { type: string, enum: [PENDING, RUNNING, SUCCESS, PARTIAL, FAILED, CANCELLED] }
        total_cashiers: { type: integer }
        success_count: { type: integer }
        awaiting_restart_count: { type: integer }
        failed_count: { type: integer }
        offline_count: { type: integer }
        started_at: { type: string, format: date-time }
    PublicationBatch:
      type: object
      properties:
        id: { type: string, format: uuid }
        advertising_block_id: { type: string, format: uuid }
        content_snapshot_json: { type: object, description: Immutable batch snapshot }
        status: { type: string, enum: [PENDING, RUNNING, SUCCESS, PARTIAL, FAILED, CANCELLED] }
        total_cashiers: { type: integer }
    PublicationBatchDetail:
      type: object
      properties:
        batch: { $ref: '#/components/schemas/PublicationBatchSummary' }
        jobs:
          type: array
          items:
            type: object
            properties:
              cashier_id: { type: string, format: uuid }
              cashier_name: { type: string }
              cashier_ip: { type: string }
              status: { type: string, enum: [PENDING, RUNNING, SUCCESS, PUBLISHED_AWAITING_RESTART, FAILED, OFFLINE] }
              attempt_count: { type: integer }
              error_message: { type: string, nullable: true }
...